@@ -5,15 +5,20 @@
 //  Copyright © 2021 Brook Street Games. All rights reserved.
 //
 
-import XCTest
+import Testing
+import UIKit
 @testable import BSGToolbarView
 
-final class ToolbarViewTests: XCTestCase {
+@MainActor
+struct ToolbarViewTests {
+    
+    // MARK: - Properties -
     
     private var basicToolbarView: ToolbarView!
     
-    override func setUp() {
-        
+    // MARK: - Initializers -
+    
+    init() {
         let tools = ["doc", "folder", "highlighter", "trash"].map { Tool(id: $0, image: UIImage(systemName: $0)) }
         basicToolbarView = ToolbarView(tools: tools)
     }
@@ -24,30 +29,28 @@ final class ToolbarViewTests: XCTestCase {
 extension ToolbarViewTests {
     
     ///
-    /// Tests initializing *ToolbarView* with a frame.
+    /// Test initializing with a frame.
     ///
-    func testFrameInitializer() {
-        
+    @Test private func testFrameInitializer() {
         let tools = ["doc", "folder", "highlighter", "trash"].map { Tool(id: $0, image: UIImage(systemName: $0)) }
         let toolbarView = ToolbarView(frame: CGRect(x: 0, y: 0, width: 320, height: 40))
         toolbarView.tools = tools
         
-        XCTAssertEqual(toolbarView.tools.count, 4)
-        XCTAssertEqual(toolbarView.activeTools.count, 0)
-        XCTAssertEqual(toolbarView.frame, CGRect(x: 0, y: 0, width: 320, height: 40))
+        #expect(toolbarView.tools.count == 4)
+        #expect(toolbarView.activeTools.count == 0)
+        #expect(toolbarView.frame == CGRect(x: 0, y: 0, width: 320, height: 40))
     }
     
     ///
-    /// Tests initializing *ToolbarView* with an array of tools.
+    /// Test initializing with an array of tools.
     ///
-    func testToolInitializer() {
-        
+    @Test private func testToolInitializer() {
         let tools = ["doc", "folder", "highlighter", "trash"].map { Tool(id: $0, image: UIImage(systemName: $0)) }
         let toolbarView = ToolbarView(tools: tools)
         
-        XCTAssertEqual(toolbarView.tools.count, 4)
-        XCTAssertEqual(toolbarView.activeTools.count, 0)
-        XCTAssertEqual(toolbarView.frame, CGRect.zero)
+        #expect(toolbarView.tools.count == 4)
+        #expect(toolbarView.activeTools.count == 0)
+        #expect(toolbarView.frame == CGRect.zero)
     }
 }
 
@@ -56,23 +59,21 @@ extension ToolbarViewTests {
 extension ToolbarViewTests {
     
     ///
-    /// Tests activating a tool.
+    /// Test activating a tool.
     ///
-    func testActivation() {
-        
-        XCTAssertTrue(basicToolbarView.activateTool(withIdentifier: "doc"))
-        XCTAssertFalse(basicToolbarView.activateTool(withIdentifier: "fakeTool"))
+    @Test private func testActivation() {
+        #expect(basicToolbarView.activateTool(withIdentifier: "doc"))
+        #expect(!basicToolbarView.activateTool(withIdentifier: "fakeTool"))
     }
     
     ///
-    /// Tests deactivating a tool.
+    /// Test deactivating a tool.
     ///
-    func testDeactivation() {
-        
+    @Test private func testDeactivation() {
         basicToolbarView.activateTool(withIdentifier: "doc")
         basicToolbarView.deactivateTool(withIdentifier: "doc")
         
-        XCTAssertEqual(basicToolbarView.activeTools.count, 0)
+        #expect(basicToolbarView.activeTools.count == 0)
     }
 }
 
@@ -81,43 +82,40 @@ extension ToolbarViewTests {
 extension ToolbarViewTests {
     
     ///
-    /// Tests the *.single* value of *selectionMode*.
+    /// Test the *.single* value of *selectionMode*.
     ///
-    func testSingleSelection() {
-        
+    @Test private func testSingleSelection() {
         basicToolbarView.selectionMode = .single
         basicToolbarView.activateTool(withIdentifier: "doc")
         basicToolbarView.activateTool(withIdentifier: "folder")
-        XCTAssertEqual(basicToolbarView.activeTools.map({ $0.id }), ["folder"])
+        #expect(basicToolbarView.activeTools.map({ $0.id }) == ["folder"])
         
         basicToolbarView.activateTool(withIdentifier: "trash")
-        XCTAssertEqual(basicToolbarView.activeTools.map({ $0.id }), ["trash"])
+        #expect(basicToolbarView.activeTools.map({ $0.id }) == ["trash"])
     }
     
     ///
-    /// Tests the *.singleLock* value of *selectionMode*.
+    /// Test the *.singleLock* value of *selectionMode*.
     ///
-    func testSingleLockSelection() {
-        
+    @Test private func testSingleLockSelection() {
         basicToolbarView.selectionMode = .singleLock
         basicToolbarView.activateTool(withIdentifier: "doc")
         
-        XCTAssertEqual(basicToolbarView.activeTools.map({ $0.id }), ["doc"])
-        XCTAssertFalse(basicToolbarView.activateTool(withIdentifier: "folder"))
-        XCTAssertEqual(basicToolbarView.activeTools.map({ $0.id }), ["doc"])
+        #expect(basicToolbarView.activeTools.map({ $0.id }) == ["doc"])
+        #expect(!basicToolbarView.activateTool(withIdentifier: "folder"))
+        #expect(basicToolbarView.activeTools.map({ $0.id }) == ["doc"])
     }
     
     ///
-    /// Tests the *.multiple* value of *selectionMode*.
+    /// Test the *.multiple* value of *selectionMode*.
     ///
-    func testMultipleSelection() {
-        
+    @Test private func testMultipleSelection() {
         basicToolbarView.selectionMode = .multiple
         basicToolbarView.activateTool(withIdentifier: "doc")
         basicToolbarView.activateTool(withIdentifier: "folder")
         basicToolbarView.activateTool(withIdentifier: "highlighter")
         basicToolbarView.activateTool(withIdentifier: "trash")
         
-        XCTAssertEqual(basicToolbarView.activeTools.count, 4)
+        #expect(basicToolbarView.activeTools.count == 4)
     }
 }
